[settings] Added hacky update interval setting.
This does a really hacky settings updater. Every time someone updates
the timer, this will restart the location checker, so it's possible
someone might need to switch settings forward then backward in order to
get settings to work properly for longer intervals (like 90min or so).

There's probably a better UX for this, but I'm not sure how that would
look.