[notification] Add URL intent for PurpleAir.
The to-do for this is to add a button so it's a bit more intuitive
(clicking the notification should ideally just take you to the app).

This will center the app around your last location automagically and with
LRAPA already turned on.

Next up:
-- A button.
-- A button in the main app that opens the same thing (too lazy right
now to read the IBind documentation).