[poller] Exp backoff. Lower power location.
Uses lower power and more frequent location updates (about once a
minute. Might revise this later).

AQI polling is now done in a simple state machine with error handling
done in an exponential backoff factor starting at 30s and increasing by
a factor of 1.5 (a successful error code is interpreted in the
RetryPolicy as a "successful" transaction when using Volley, so this had
to be done manually).

Probably should write some tests for this. Behavior is unchanged.