[sensor] Use EPA correction standard.
This does a best-effort EPA style correction of PM2.5 for PurpleAir
sensors, taking into account relative humidity.

This does not attempt to either drop or average Channel A or B from the
sensors (as is recommended by the EPA paper) as the json query API isn't
well documented. I am still in the process of figuring out how to do
this. I am quite certain that I am getting a CF=1 value from the
sensors, though.